Updates to charts, some visual improvements, change in way data moves to charts
