+++ app/assets/stylesheets/NavBar.scss
@@ -17,3 +17,15 @@
 .rounded{
   border-radius: 10rem
 }
+
+.grid-container{
+  margin: 1rem 5rem 1rem 5rem
+}
+
+.callout{
+  border-color: #97F4E1;
+}
+
+select{
+  border-radius: 5rem;
+}
