+++ app/javascript/react/components/FullTable.js
@@ -1,11 +1,9 @@
 import * as React from "react";
 import { useEffect, useState } from "react"
 import { Chart } from "react-google-charts"
-import PrepareTable from "./PrepareTable"
 
 const FullTable = props =>{
-  const [tableData, setTableData] = useState(["table"])
-  const [dataObject, setDataObject] = useState({})
+  const [dataObject, setDataObject] = useState("{}")
   const [dataFetched, setDataFetched] = useState(false)
 
   const fetchDataObject = async () => {
@@ -28,34 +26,41 @@ const FullTable = props =>{
   useEffect(() => {
     fetchDataObject()
   }, [])
-debugger
-  const prepareData = () =>{
-    return(
-    <div>
-    <PrepareTable dataObject={dataObject} setDataObject={setDataObject}/>
-    </div>)
-  }
- 
-  if((dataFetched == true) && (tableData[0] === "table")){
-    prepareData()
-  }
 
-  // return(
-  //   <div>
-  //     <PrepareTable dataObject={dataObject} setDataObject={setDataObject}/>
-  //     <h1>Test</h1>
-  //     <Chart
-  //     width={'auto'}
-  //     height={'auto'}
-  //     chartType="Table"
-  //     loader={<div>Loading Chart</div>}
-  //     data={tableData}
-  //     options={{
-  //       showRowNumber: false,
-  //     }}
-  //     rootProps={{ 'data-testid': '1' }}
-  //   />
-  // </div>
-  // )
+
+const dataHeads = [
+  { type: 'string', label: 'Date' },
+  { type: 'string', label: 'TNSS' },
+  { type: 'string', label: 'Tree Pollen Level' },
+  { type: 'string', label: 'Tree Pollen Intensity'},
+  { type: 'string', label: 'Grass Pollen Level' },
+  { type: 'string', label: 'Grass Pollen Intensity'},
+  { type: 'string', label: 'Ragweed Pollen Level' },
+  { type: 'string', label: 'Ragweed Pollen Intensity'},
+  { type: 'string', label: 'Mold Level' },
+  { type: 'string', label: 'Mold Intensity'},
+  { type: 'string', label: 'Dust and Dander Level' },
+  { type: 'string', label: 'Dust and Dander Intensity'},
+]
+  return(
+    <div>
+      <Chart
+      width={'auto'}
+      height={'auto'}
+      chartType="Table"
+      loader={<div>Loading Chart</div>}
+      data={[
+        dataHeads,
+        ["12/1/21", "10", "1", "Low", "1", "Low", "1", "Low", "1", "Low", 
+        "1", "Low",]
+
+      ]}
+      options={{
+        showRowNumber: false,
+      }}
+      rootProps={{ 'data-testid': '1' }}
+    />
+  </div>
+  )
 }
 export default FullTable
