--- db/migrate/20211124203604_create_zip_codes.rb
@@ -1,9 +1,0 @@
-class CreateZipCodes < ActiveRecord::Migration[6.1]
-  def change
-    create_table :zip_codes do |t|
-      t.string :zip_code, null: false, limit: 5
-
-      t.timestamps
-    end
-  end
-end
