--- db/migrate/20211124204503_add_ref_to_user.rb
@@ -1,5 +1,0 @@
-class AddRefToUser < ActiveRecord::Migration[6.1]
-  def change
-    add_reference :users, :zip_code, foreign_key: true, null: false
-  end
-end
