+++ db/seeds.rb
@@ -6,6 +6,24 @@
 #   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
 #   Character.create(name: 'Luke', movie: movies.first)
 
+codes = ZipCode.create([
+  { zip_code: "02108" },
+  { zip_code: "02109" },
+  { zip_code: "02110" },
+  { zip_code: "02113" },
+  { zip_code: "02114" },
+  { zip_code: "02141" },
+  { zip_code: "02142" },
+  { zip_code: "02138" },
+  { zip_code: "02139" },
+  { zip_code: "02140" },
+  { zip_code: "02143" },
+  { zip_code: "02144" },
+  { zip_code: "02145" },
+  { zip_code: "02210" },
+  { zip_code: "02215" },
+])
+
 users = User.create([{ 
   email: "test@test.com",
   password: "tester",
@@ -99,22 +117,5 @@ allergens = AllergenReport.create([{
   dust_dander_category: "Low",
   created_at: Date.new(2021, 11, 27),
   zip_code_id: 1
-},
-
-codes = ZipCode.create([
-  { zip_code: "02108" },
-  { zip_code: "02109" },
-  { zip_code: "02110" },
-  { zip_code: "02113" },
-  { zip_code: "02114" },
-  { zip_code: "02141" },
-  { zip_code: "02142" },
-  { zip_code: "02138" },
-  { zip_code: "02139" },
-  { zip_code: "02140" },
-  { zip_code: "02143" },
-  { zip_code: "02144" },
-  { zip_code: "02145" },
-  { zip_code: "02210" },
-  { zip_code: "02215" },
+}
 ])
